api/pyproject.toml
[project]
name = "api"
version = "0.0.1"
description = "RAG BOT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "fastapi>=0.115.6",
    "langchain-chroma>=0.1.4",
    "langchain-community>=0.3.13",
    "langchain-core>=0.3.28",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.2.2",
    "langchain-text-splitters>=0.3.4",
    "onnxruntime>=1.20.1",
    "pycodestyle>=2.12.1",
    "pydantic-settings>=2.7.0",
    "pypdf>=5.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sseclient>=0.0.27",
    "streamlit>=1.41.1",
]

[tool.pycodestyle]
ignore = "E501"
max-line-length = 120
